+++ setWavePeriod.cpp
@@ -1,6 +1,6 @@
 #include "synth_os.h"
 #include "updateOutput.h"
 
-void UpdateOutput::setWavePeriod(float newSamplesInPeriod){
+void UpdateOutput::setWavePeriod(int newSamplesInPeriod){
 	samplesInPeriod = newSamplesInPeriod;		//sets private attribute samples in period
 }
